+++ src/solvers/dgmulti/types.jl
@@ -76,25 +76,56 @@ function DGMulti(element_type::AbstractElemShape,
   return DG(rd, nothing #= mortar =#, surface_integral, volume_integral)
 end
 
-
-# now that DGMulti is defined, we can define constructors for VertexMappedMesh which use dg::DGMulti
+# now that `DGMulti` is defined, we can define constructors for `DGMultiMesh` which use `dg::DGMulti`
 """
-    VertexMappedMesh(vertex_coordinates, EToV, dg::DGMulti;
-                     is_on_boundary = nothing,
-                     is_periodic::NTuple{NDIMS, Bool} = ntuple(_->false, NDIMS)) where {NDIMS, Tv}
-
-Constructor which uses `dg::DGMulti` instead of `rd::RefElemData`.
-"""
-VertexMappedMesh(vertex_coordinates, EToV, dg::DGMulti; kwargs...) =
-  VertexMappedMesh(vertex_coordinates, EToV, dg.basis; kwargs...)
-
-"""
-    VertexMappedMesh(triangulateIO, dg::DGMulti, boundary_dict::Dict{Symbol, Int})
+  DGMultiMesh(vertex_coordinates, EToV, dg::DGMulti;
+              is_on_boundary = nothing,
+              periodicity::NTuple{NDIMS, Bool} = ntuple(_->false, NDIMS)) where {NDIMS, Tv}
 
 Constructor which uses `dg::DGMulti` instead of `rd::RefElemData`.
 """
-VertexMappedMesh(triangulateIO, dg::DGMulti, boundary_dict::Dict{Symbol, Int}) =
-  VertexMappedMesh(triangulateIO, dg.basis, boundary_dict)
+DGMultiMesh(vertex_coordinates, EToV, dg::DGMulti; kwargs...) =
+  DGMultiMesh(vertex_coordinates, EToV, dg.basis; kwargs...)
+
+"""
+  DGMultiMesh(triangulateIO, dg::DGMulti, boundary_dict::Dict{Symbol, Int})
+
+Constructor which uses `dg::DGMulti` instead of `rd::RefElemData`.
+"""
+DGMultiMesh(triangulateIO, dg::DGMulti, boundary_dict::Dict{Symbol, Int}) =
+  DGMultiMesh(triangulateIO, dg.basis, boundary_dict)
+
+"""
+    DGMultiMesh(dg::DGMulti; cells_per_dimension,
+                coordinates_min=(-1.0, -1.0), coordinates_max=(1.0, 1.0),
+                is_on_boundary=nothing,
+                periodicity=ntuple(_ -> false, NDIMS))
+
+Constructs a Cartesian [`DGMultiMesh`](@ref) with element type `dg.basis.elementType`. The domain is
+the tensor product of the intervals `[coordinates_min[i], coordinates_max[i]]`.
+- `is_on_boundary` specifies boundary using a `Dict{Symbol, <:Function}`
+- `periodicity` is a tuple of `Bool`s specifying periodicity = `true`/`false` in the (x,y,z) direction.
+"""
+function DGMultiMesh(dg::DGMulti{NDIMS}; cells_per_dimension,
+                     coordinates_min=ntuple(_ -> -one(real(dg)), NDIMS),
+                     coordinates_max=ntuple(_ -> one(real(dg)), NDIMS),
+                     is_on_boundary=nothing,
+                     periodicity=ntuple(_ -> false, NDIMS), kwargs...) where {NDIMS}
+
+  if haskey(kwargs, :is_periodic)
+    # TODO: DGMulti. Deprecate `is_periodic` in version 0.5
+    Base.depwarn("keyword argument `is_periodic` is now `periodicity`.", :DGMultiMesh)
+    periodicity=kwargs[:is_periodic]
+  end
+
+  vertex_coordinates, EToV = StartUpDG.uniform_mesh(dg.basis.elementType, cells_per_dimension...)
+  domain_lengths = coordinates_max .- coordinates_min
+  for i in 1:NDIMS
+    @. vertex_coordinates[i] = 0.5 * (vertex_coordinates[i] + 1) * domain_lengths[i] + coordinates_min[i]
+  end
+
+  return DGMultiMesh(vertex_coordinates, EToV, dg, is_on_boundary=is_on_boundary, periodicity=periodicity)
+end
 
 # Todo: DGMulti. Add traits for dispatch on affine/curved meshes here.